requirements.txt
# Streamlit Cloud deployment requirements
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=9.5.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0